lidar/requirements.txt
contourpy==1.2.0
cycler==0.12.1
fonttools==4.48.1
keyboard==0.13.5
kiwisolver==1.4.5
llvmlite==0.41.1
matplotlib==3.7.4
numba==0.58.1
numpy==1.24.4
opencv-python==4.7.0.72
packaging==23.2
pillow==10.2.0
pyparsing==3.1.1
pyserial==3.5
python-dateutil==2.8.2
rplidar-roboticia==0.9.5
scipy==1.10.1
six==1.16.0
